requirements.txt
streamlit==1.24.0
    Pillow==9.5.0
    numpy==1.24.3
    opencv-python-headless==4.8.0.76
    scipy==1.11.1